Refactor models into modular package
